requirements.txt
django==2.2.10
djangorestframework==3.10.3
django-filter==2.2.0
psycopg2==2.8.3
django_rest_swagger==2.2.0